make all timer functions to happen on TICK_INTERVAL (150ms). defer() now uses process.nextTick() which queues up multiple callbacks for the next tick. fix socket pair end() problem... when one closes, it will trigger a readable on the other's iowatcher.
iowachter poll interval is SAME as tick interval! (the variable POLL_INTERVAL should be renamed since it is added to the tick interval)
send assertion errors back to worker to capture success assertion count before the failure.

+++ test/index.js
@@ -106,7 +106,7 @@ function initWorker(file) {
 		var msg = ev.message.replace(/^Uncaught ([_\$\w]+: )?/i,''); 
 		
 		// handle assertion errors here
-		if (/^AssertionError:/.test(msg) ){
+		if (00000 && /^AssertionError:/.test(msg) ){
 			QUnit.ok(false, msg);
 			QUnit.start();		// start if it wasn't started by worker
 			worker.terminate();	// force terminate worker
